extract MessageInsertExtension to its own file
